Remove delay-based backpressure in favor of explicit queue limits (#1515)
Right now, we have a backpressure delay that's a function of `(job
count, bytes in flight)`.

The actual queue length is set implicitly by the shape of that function
(e.g. its gain) and the actual downstairs IO time. If a downstairs slows
down due to load, then the queue length has to grow to compensate!

Having an implicitly-defined queue length is tricky, because it's not
well controlled. Queue length also affects latency in subtle ways, so
directly controlling the queue length would be helpful.

This PR removes the delay-based backpressure implementation in favor of
a simple pair of semaphores: there are a certain number of job and byte
permits available, and the `Guest` has to acquire them before sending a
job to the `Upstairs`.

In other words, writes will be accepted as fast as possible _until_ we
run out of permits; we will then shift to a one-in, one-out operation
mode where jobs have to be completed by the downstairs before a new job
is submitted. The maximum queue length (either in jobs or bytes) is well
known and set by global constants.

Architecturally, we replace the `BackpressureGuard` with a new
`IOLimitGuard`, which is claimed by the `Guest` and stored in relevant
`BlockOp` variants (then moved into the `DownstairsIO`). It still uses
RAII to automatically release permits when dropped, and we still
manually drop it early (once a job is complete on a particular
downstairs).