Reset completed work Downstairs on a `Barrier` operation (#1601)
A `Barrier` operation should also clear the Downstair's list of complete
jobs, because it's guaranteed to be the sole dependency for all future
jobs. The current implementation is correct but unnecessarily
conservative, and will accumulate arbitrary amounts of complete jobs if
IO is constantly streaming without a flush.

I believe this was an oversight in #1494 – we handled it correctly in
`test::complete`, but not in the actual non-test code!