Generic read test for crutest
Made a generic read test for crutest, just do reads (and
an occasional flush).  It makes use of the generic test that
is already present and does not choose to ever write.

This is paving the way for a read only stress test with downstairs
coming and going.

Added the ability for the crutest cli to call this new read only test.