Remove `LastFlushAck` (#1603)
Right now, if a downstairs is coming online through replay
(`ConnectionMode::Offline`), the Upstairs goes through the following
steps to finish negotiation:

- From`WaitForRegionInfo`, it sends `LastFlush`, which gives the
Downstairs a lower bound for job IDs
  - This assignment is infallible on the Downstairs side
- The upstairs enters `NegotiationState::GetLastFlush` while waiting for
a reply
- The Downstairs then sends back that same flush ID in a `LastFlushAck`
message
- Once the Upstairs receives this message, it exits negotiation and
replays all previous jobs

There are two things that are awkward about this sequence
- If someone is continuing to send IO, the Downstairs can become
ineligible for replay between sending `LastFlush` and `LastFlushAck`,
because we could exceed our "how much IO to buffer" limit. It's not
_wrong_ in the current implementation – we disconnect and restart
negotiation from the start – but having that window of fallibility is
awkward.
- The `LastFlushAck` reply does not convey much information. Message
handling in the Downstairs is infallible and will always return the same
job ID.

This PR removes the `LastFlushAck` reply and `GetLastFlush` negotiation
state; instead, replayed jobs are sent immediately following the
`LastFlush` message. We already assume that messages are reliable and
strongly ordered (thanks, TCP!), so when the Downstairs gets the first
replayed job, it **must** have already received the `LastFlush` message.

Because we're still doing rack-wide updates, I'm bumping the Crucible
message version and removing the `LastFlushAck` entirely. I _think_ this
is still fine; otherwise, we can keep that variant (but mark it as
deprecated).

(This is secretly a building block for Stateless Negotiation™, but makes
sense as a stand-alone change)