Tweaks to automatic flush (#1613)
There was previously an edge case where we'd submit automatic flushes
500ms after the last Upstairs event, _even if_ we had IO pending. We
normally don't see this in practice, because events are constantly
happening (and therefore resetting the flush timer).

However, @jmpesp saw it while testing in the Canada region; we're not
yet sure why!

This PR makes two changes: (1) do not send flushes if jobs are pending,
even if the timeout has expired, and (2) log a specific warning message
about this unusual situation.