Allow read only activation with less than three downstairs (#1608)
Allow a read only upstairs to activate with one a single downstairs 
present.

In upstairs/src/upstairs.rs, I've added a check when a downstairs
transitions to `WaitQuorum`.  If we are read-only, then we can skip
reconciliation and activate the upstairs.  If we are already active 
(and read only), then a new downstairs can go to active.

Added some tests and a bit of additional test framework to verify
an upstairs can activate with only a single downstairs ready.

This "fixes" the feature request in
https://github.com/oxidecomputer/crucible/issues/1599
and may help with https://github.com/oxidecomputer/crucible/issues/1593