Use `RangeSet` instead of tracking every complete job (#1663)
Right now, we store each completed `JobId` and search linearly through
the list when checking for dependencies. This means that a flush, which
is dependent on all previous jobs, takes `O(N^2)` time to check its
dependencies.

This PR switches to a
[`RangeSet`](https://docs.rs/rangemap/latest/rangemap/set/struct.RangeSet.html),
which provides `O(log N)` lookup. In practice, we expect it to be
`O(1)`: jobs are completed in order, so the `RangeSet` will only contain
a single contiguous range.

I did a quick audit of the `rangemap` crate, and it seems reasonable:

- No dependencies
- `proptest` based testing, Github CI
- Consistent docstrings

There are two small behavior changes:

- `completed()` now allocates, expanding the `RangeSet` into a
  `Vec<JobId>` (instead of a `&[JobId]`). It's only used in unit tests and
  our debug work-printing function, so I think that's fine.
- The output of `completed()` is now sorted. This required changing one
  test assertion.