nightly test polish (#1665)
Updates to the nightly tests and the tests it calls.

All the nightly tests have their own named directory for test logs
and for the location of the downstairs regions.
```
WORK_ROOT=${WORK_ROOT:-/tmp}
TEST_ROOT="${WORK_ROOT}/<test_name>"
```

This makes cleanup at the end easier.
In all tests, dsc now logs to this or a specified directory.

All the nightly tests now have a `REGION_ROOT`, and will create a test specific
directory inside this location.   This is where dsc or the crucible downstairs 
will create region directories. This makes it easier for the caller to direct the
tests to use a specific location instead of everyone piling in to `/var/tmp`

If a test finishes without error, it will clean up logs and regions that
were created as part of the test. If there is an error, then we leave
things behind. We clean up the contents of `REGION_ROOT` as well.

Updated `tools/test_nightly.sh` to print the name of the test it is
running, and just cleaned up the output a bit.

---------

Co-authored-by: Alan Hanson <alan@oxide.computer>