Make `ImpactedBlocks` an absolute range (#1685)
The current implementation of `ImpactedBlocks` stores an inclusive range
of `ImpactedAddr`, which are themselves `(ExtentId, BlockOffset)`
tuples. This makes life awkward, because we don't know how to step
through a range of `ImpactedBlocks` without knowing the
blocks-per-extent count: is `(ExtentId(0), BlockOffset(7))` adjacent to
`(ExtentId(1), BlockOffset(0))`? Without additional information, there's
no way of knowing!

This ambiguity adds a bunch of boilerplate, and specifically complicates
the implementation of `ActiveJobs`.

The only time when we actually _need_ the extent ID is when doing
live-repair activities, which forgo the block abstraction and care about
individual extent files. It makes much more sense to have
`ImpactedBlocks` be an absolute block range (using `BlockIndex` instead
of `BlockOffset`), then tweak the small number of sites which care about
extents. These sites are:

- `Downstairs::reserve_repair_ids_for_extent`
- `Downstairs::get_repair_ids`
- `Downstairs::check_repair_ids_for_range`
- `ActiveJobs::deps_for_repair`
- `ImpactedBlocks::extents`