Add more extents for repair tests (#1689)
The repair tests in tools/test_up.sh were not creating enough extents.
The replace-reconcile test wants enough extents to repair such that we
can start things repairing and then interrupt it before it finishes.

Fix for the flake: https://github.com/oxidecomputer/crucible/issues/1686

Also, the "echo" was out of sync with the actual command we ran, fixed
that too.