Target replacement support for VCRs with multiple sub_volumes (#1551)
Updated the volume layer code to support replacement of a target in a
VCR that has multiple sub_volumes in it.
Before this we only supported a single sub_volume.

Most of the changes here are for tests. I've added a loop around the VCR
replacement tests to create multiple sub volumes and iterate through the
test for different VCR configurations.

This relaxes the old requirement that a new VCR have the sub-volume
generation bumped. Going forward VCR changes will only verify that
generation numbers in a new VCR are equal or greater than what is in 
the current VCR. An Increase in the generation number is no longer required.

Some tests that were duplicates have been removed, and a new layer of
tests that don't bump generation numbers was added.

---------

Co-authored-by: Alan Hanson <alan@oxide.computer>