bump Progenitor to 0.10.0
my true goal is to bump Progenitor in Nexus to 0.10.0, but because Nexus
uses `crucible-agent-client` from this repo in the `nexus` crate while
_also_ using `progenitor`, the Progenitor versions in Nexus and the rev
of Crucible that Nexus depends on must be the same.

so, bump it here, then bump Crucible there. question for reviewers here:
do i want to also bump Crucible in Propolis as well? meaning i'd merge
this, then repeat this against Propolis, then update both in Omicron?

Progenitor 0.10.0 is interesting because it adds a new `api-version`
header to requests with whatever version is in the OpenAPI spec. this is
suitable for, among other things, use with Dropshot API versioning!

i did check that Dropshot acts reasonably when an unexpected
`api-version` is present, so this doesn't necessitate changes to
HTTP servers in Crucible.

i'd like to untangle Nexus from Crucible's Progenitor version. i *think*
we can use omicron-common here, export a `ProgenitorOperationRetry`, and
use the `crucible_agent_client::ProgenitorOperationRetry`. this would be
marginally better than the status quo of expecting Nexus and Crucible to
agree on `progenitor` version so they don't conflict on the
`progenitor_client::Error` type.