[agent-client] re-export retry type to untangle Progenitor versioning
this is the first of two steps in not having to make changes like
https://github.com/oxidecomputer/omicron/pull/7333, and i'm about to
make in service of Progenitor 0.10.0.

Omicron needs to have the same Progenitor version as Crucible (and
interestingly not other dependencies) because Nexus uses a
`ProgenitorOperationRetry` type to ... retry some kinds of Progenitor
errors. critically, this same retry type is used for Omicron-internal
clients with Omicron's Progenitor *and* to wrap clients that Omicron
uses from external dependencies like `crucible-agent-client`. then, if
Crucible's Progenitor version is different, Omicron fails to build
because the retry wrapper uses Omicron-version Progentior when it is
given a Crucible-version Progenitor instead and the types are not
identical.

instead, if we depend on omicron-common in crucible-agent-client, export
*our* version of ProgenitorOperationRetry defined in terms of *our*
version of Progenitor, users of crucible-agent-client that want to retry
can use our type instead.

if someone wants to abstract over all clients' ProgenitorOperationRetry,
like if there was a list of mixed sled-agent and crucible operations to
work through, we'd have this issue again.