DTrace updates (#1715)
Updated the upstairs DTrace struct to include the UUID of the upstairs
itself. This is helpful if a disk has more than one sub-volume and you
want to determine which output matches what.

Renamed the struct used by DTrace from "Arg" to "DtraceInfo".

Changed many of the dtrace scripts to also print the UUID of the
upstairs.

Fixed sled_upstairs_info.d to print the new 3 character state string.