Make Downstairs stoppable (#1730)
Right now, `start_downstairs` returns a `RunningDownstairs` handle that
includes a `JoinHandle` for the network listener task. It also spawns
several other tasks:

- The task running `Downstairs::run`
- The repair server
- An optional Oximeter server
- Per-connection tasks (spawned on demand)

The lifetime of these tasks is unclear, and it's hard to stop them all.
Aborting the network listener task doesn't actually stop any of the
other tasks!

When writing unit tests (not currently pushed), I wanted to be able to
stop and restart an entire Downstairs task group. This would make it
easier to write integration tests in Rust, instead of using separate
processes orchestrated by Bash.

This PR renames `RunningDownstairs` to `DownstairsClient`, and gives it
the power to halt the full set of Downstairs tasks. In addition,
`start_downstairs` is moved to `DownstairsClient::spawn`, and takes a
new `DownstairsClientSettings` object (instead of 7 individual
arguments).

Finally, certificate info is moved into a
`Option<DownstairsClientCerts>` object, providing all-or-none behavior;
previously, we could panic if only some certificate were provided. This
is an internal change; `CrucibleOpts` preserves the old behavior,
because it has been serialized to disk.