Use data-bearing enum variants pattern in negotiation (#1727)
During negotiation, there are certain pieces of data which are only
relevant in particular states:

- Before connecting, the client has a `oneshot::Sender<()>` which kicks
off the connection when triggered
- When waiting for a quorum, each client stores a `RegionMetadata`, to
be used when computing reconciliation

Previously, these two pieces of data were always present in the
`DownstairsClient` object, and were checked (or unwrapped) at the
relevant points in negotiation.

This PR moves them into the relevant `NegotiationState` variants, so
they are _guaranteed_ to be present at the right times.

Unfortunately, `NegotiationState` is also used (as a member of
`DsState`) in serialization, for both OpenAPI calls and DTrace probes.
We can't serialize or log a oneshot sender (or region metadata), which
is a problem!

The **vast majority** of this PR is separating the
serialization-friendly subset of data (`DsState` and `NegotiationState`)
from the data-bearing variants (new `DsStateData` and
`NegotiationStateData`).

There's also one subtlety: the PR adds a new
`NegotiationState::WaitConnect(oneshot::Sender<()>)`. Previously,
`NegotiationState::Start` represented both "waiting for connection
oneshot to fire" and "waiting for connection to the Downstairs"
(depending on whether `ClientTaskHandle::client_connect_tx` was present
or not). This change makes those two sub-states distinct.