Improve `ClientIoTask` start logic (#1731)
In the Crucible upstairs, we stop the client IO task for various
reasons, which fall into two broad categories:

- Something has gone wrong, and we need to recover with a fresh
connection. In this case, it's wise to sleep for a little while; we
don't want to spam connections to an offline Downstairs as fast as
possible. This is controlled by a `delay: bool`, flag which adds a
10-second delay before reconnecting.
- We are deliberately stopping the upstairs. In this case, we don't want
to try connecting at all until the upstairs goes active again. This is
controlled by an `Option<oneshot::Receiver<()>>`, which is awaited
before connecting if present.

In the current codebase, when the upstairs is disabled, we restart the
Client IO task with **both** of these measures in place: the Client IO
task waits for a message on a oneshot, _then_ sleeps for 10 seconds
before connecting. This is awkward, because it means that a `GoActive`
request is blocked for 10 seconds.

We could just fix this issue, but I realized that the flags are mutually
exclusive: we should only ever start an IO task in one or the other
mode. As such, we can use a new `enum ClientConnectDelay` to represent
the choice. This ends up being a nice simplification!