Add back job delays with the --lossy flag
Somewhere in the past we lost the ability for the --lossy flag
in the downstairs to make some jobs take longer.  The test
test_repair.sh was counting on this behavior.  Things still generally
behaved the same, but the volume of repair that the test would do
was greatly reduced.  This change adds back the delay and will result
in whatever downstairs has the lossy flag falling much further behind
the other downstairs, which is what we want for that test.