Snapshots existing already are ok! (#1759)
Previously, the Downstairs would consider a snapshot existing already as
an error, and unfortunately the Upstairs would continuously retry in the
face of these errors, leading to an indefinite retry loop if any
previous flush (with snapshot) succeeded.

This commit removes the SnapshotExistsAlready error variant, and returns
Ok if the snapshot exists already. This is ok as long as:

- job dependency lists are honoured, meaning no writes can land that
would make each downstairs inconsistent

- snapshot names are random enough that collisions are rare, and
therefore an existing snapshot can be treated as the previous attempt
succeeding

Fixes #1758.