Offline downstairs go to faulted when LR starts (#1777)
If a downstairs goes offline before, or during a Live Repair, we must
not allow any replay of IO to that downstairs. The Live Repair process
requires all downstairs to go through specific steps at the same time to
be sure all three will agree, if there are repair IOs outstanding, the
will be blocked while the missing downstairs is away, and all IOs will
come to a halt backing up behind them.

To handle this, we move any downstairs that is Offline to faulted and
have it come back through LiveRepair once it returns.

Added a new tests to crutest that will allow IO to keep flowing even if
there is an error. It also does a better job of tracking what we expect
to have written to each block.

Added two tests to check that offline downstairs are handled correctly
during Live Repair situations.

---------

Co-authored-by: Alan Hanson <alan@oxide.computer>