Add two repair tests to agent-antagonist
Discussing possible causes for the zeroes appearing at the beginning of
an extent file (#1788), one theory that came up was that the Extent was
being repaired, and those zeroes came from reading the Extent from the
repair API, not from the disk itself: only _one_ Region's Extents were
bad, not all of them.

In order to test this theory, add two new tests, both of which will
contact an Agent to get a Region's details, then use the repair API to
read Extent data files and either look for a block of zeroes at the
beginning or write that to a temporary file and use the new
`Extent::validate` routine.