Enforce write order of blocks and dirty bit (#1798)
Crucible's **raw file extent** relies stronger-than-POSIX requirements
on write ordering. Specifically, we expect that if we submit a series of
writes through a single file descriptor, they will land on disk
atomically and in the order that they were submitted; we should not see
tearing or reordering.

This is guaranteed by ZFS, with one caveat: performing an `fsync` or
`syncfs` on a set of writes which includes **multiple writes to the same
location** will write the **most recent data** when persisting earlier
writes.

To understand why, consider the following sequence of writes:

```
| number | block 0 | block 1 |
|--------|---------|---------|
| 0      |  hello  |         |
| 1      |         | world   |
| 2      |  aahhh  |         |
```

If we call `fsync`, writes are persisted to the **ZIL** in order as
`TX_WRITE` objects. However, when persisting a write, ZFS goes to the
DMU to decide what to log in the ZIL. The DMU will return the **current
state** of the file. Persisting write 0 to disk will record `aahhh`
instead of `hello`.

This means that if we pull power in between persisting writes 1 and 2,
the ZIL replay will reconstruct `| aahhh | world |`


--------------------------------------------------------------------------------

This is not a new discovery (to me, at least). When designing the raw
file extent, I added special logic to add an `fsync` call before
overwriting a context slot, to prevent this very issue.

However, there are two cases which we were not handling correctly.

# Dirty bit handling

Each raw extent file contains two "context slots" for each block. We
have two context slots so that we can write an inactive slot, then write
block data; if we lose power in between, then the previous (active)
context slot allows us to decrypt the block data.

We must be able to tell which context slot is active. To do so, we write
a hash of the block data to the context slot. On startup, we can hash
each block and compare against the two context slots; whichever one has
a matching hash is active.

Unfortunately, this is slow. As an optimization, when flushing the file,
we write an "active slot array", which records which slot is active for
each block. This active slot array is only valid when the `dirty` flag
in the `OnDiskMeta` object is `false`; this flag is set before any
write, and cleared in the same operation that writes the active slot
array.

This is all well and good. Now, consider the following sequence of
writes

```
| number | block 0 | slot A | slot B | active slot array | dirty |
|--------|---------|--------|--------|-------------------|-------|
| 0      |         |        |        |                   | true  |
| 1      |         |  ctx   |        |                   |       |
| 2      | data    |        |        |                   |       |
| 3      |         |        |        | block 0 -> slot A | false |
```

When persisting these writes to the ZIL, write 0 will persist `dirty =
false` (because that's the current state of the file). This is bad,
because a failure after this point (but before write 3) means that we'll
assume the active slot array is correct, which may not be the case!

This leads to cases where the active slot array is used (because `dirty
= false`) but points to the wrong context slot. Here's an example from
#1788:
```
bs:512 bytes  bc:131072  dirty:false gen:2 flush_number:  5063 ext_ver:2 bonus_sync:157 defrag:0
Error at block  56624:  slot A [selected]: Err(SlotHashMismatch), flush id: 5057 | slot B [deselected]: Ok(HashMatch), flush id: 5063
```

The fix here is to add a `fsync` call after writing `dirty = true`.

# Repeated writes to the same block

This one is even simpler:
```
| number | block 0 | slot A | slot B |
|--------|---------|--------|--------|
| 0      |         |  ctx 1 |        |
| 1      | data 1  |        |        |
| 2      |         |        | ctx 2  |
| 3      | data 2  |        |        |
```

When persisting data to disk, write 1 will persist `data 2`. If we fail
between writes 1 and 2, then we will have data on disk for which we have
no valid context. This data cannot be decrypted, which is Bad News.

@JustinAzoff saw this kind of failure mode when running Jepsen tests
([matrix](https://matrix.to/#/!sHZivBWjdefFxgKtLU:oxide.computer/$fLqhwDlZRu_Bm1s2t4U3nT6Xf8p1r7_FyKgmI6RuUGo?via=oxide.computer&via=unix.house&via=matrix.org)):
```
bs:512 bytes  bc:   100  dirty: true gen:1 flush_number:    13 ext_ver:2 bonus_sync:17 defrag:0
Error at block 11:  slot A: SlotHashMismatch, flush id: 3 | slot B: SlotHashMismatch, flush id: 13
```

The fix here is similar: we add an `fsync` call before overwriting a
block.


--------------------------------------------------------------------------------

This PR makes a few other changes:

- Extract a more generic `BlockBitArray` from `ActiveSlotArray`, then
  use it for both `ActiveSlotArray` and to track which blocks are dirty
- Add debug info to `OnDiskDownstairsBlockContext` and `OnDiskMeta`
- This takes advantage of the fact that their slots were oversized (e.g.
  we were only using 45 of 48 bytes for `OnDiskDownstairsBlockContext`)
- We can use this info for more sophisticated debugging of failures like
  this
  
Hopefully fixes #1788