Remove cargo nextest from github workflow, out of space (#1846)
The github provided ubuntu image/host/container has a limited amount of
space, and we have apparently hit that limit. We could figure it out and
update it, but this test is already a duplicate of what we do in
buildomat, so I just pulled out the test part, so now it just does cargo
fmt/build/check.

---------

Co-authored-by: Alan Hanson <alan@oxide.computer>